requirements.txt
jax
jaxlib
pax3
chex
jmp
